requirements.txt
fastapi
pymongo
argon2-cffi
rich
uvicorn[standard]